fix(ci): HWPX 되감김 승격을 쪽 하단 넘침으로 좁혀 pr-1674 36쪽 회귀를 막는다
internal test의 line_index(u32)를 usize로 맞춰 컴파일을 고치고,
쪽 중간 키 큰 문단의 저장 되감김은 쪽 경계로 올리지 않는다.

refs #6063

+++ tests/cases/issue_6063_hwpx_midpara_vpos_rewind_body_overflow.rs
@@ -38,7 +38,7 @@ fn collect_body_text_lines(node: &RenderNode, lines: &mut Vec<(usize, usize, f64
     if in_body {
         if let RenderNodeType::TextLine(tl) = &node.node_type {
             if let (Some(pi), Some(li)) = (tl.para_index, tl.line_index) {
-                lines.push((pi, li, node.bbox.y + node.bbox.height));
+                lines.push((pi, li as usize, node.bbox.y + node.bbox.height));
             }
         }
     }
